requirements.txt
prometheus_client
requests
aiohttp
pyyaml
